// Fontend/TNT.N8.Web/TNT.N8.Web/ClientApp/src/app/order/models/customer-order-detail.model.ts
export class CustomerOrderExtension {
  id: string;
  attributeConfigurationId: string | null;
  objectId: string | null;
  objectType: string;
  value: string;
  dataType: number;
  createdById: string;
  createdDate: string;
  updatedById: string;
  updatedDate: string;
  tenantId: string;
  orderDetailId: string;
  attrName: string;
  servicePacketMappingOptionsId: string;
}